Raise DCSError when communication with DCS fails
previously such an exception was raised only from the `get_cluster()`
method, and now we will to do the same from the `update_leader()` and
`attempt_to_acquire_leader()` methods. These methods influence Postgres
promotion and demotion and we want to make a difference between
different kinds of failures. Specifically, if calls have failed because
DCS isn't accessible or due to a timeout.

In the future, this exception could be handled differently.