HA.is_leader() could be reset if DCS suddenly available for a moment
If it happens, the check_failsafe_topology() isn't triggered and
postgres is demoted.

Instead we will simply rely on the fact that postgres running as a
primary (pg_is_in_recovery() = false) to trigger the
check_failsafe_topology() and modify the `POST /failsafe` endpoint
that it doesn't return a success if for some reason other node is
running as a primary.