Introduce numsync_confimed
the number of nodes from synchronous_standby_names that are guaranteed
to be synchronous. The guaranty is verifyed by comparing replay LSN
against the primary.

if the value of `numsync_confimed` is smaller than `numsync` the quorum
requirements is increased by the difference.

In case if `numsync_confimed` is 0 (could happen after failover/restart)
recalculate it from `voters & sync` intersection.