Register secondaries in pg_dist_node
1. All nodes with role == 'replica' and state == 'running' are
   registered. In case is state isn't running the node is removed.
2. In case of failover/switchover we always first update the primary
3. When switching to a registered secondary we call citus_update_node()
   three times: rename primary to primary-demoted, put the primary name
   to a promoted secondary row and put secondary name to the primary row

State transitions are produced by the transition() method.
First of all the method makes sure that the actual primary is registered
in the metadata. In case if for a given group the primary didn't change,
the method registers new secondaries and removes secondaries that are
gone. It preferes to use citus_update_node() UDF to replace gone
secondaries with added.

Communication protocol between primary nodes remains the same and all
old features work without any changes.